+++ CHANGES.rst
@@ -5,7 +5,7 @@ Changelog of lizard-damage
 1.2 (unreleased)
 ----------------
 
-- Nothing changed yet.
+- Make risk maps work with refactored result collector.
 
 
 1.1 (2014-08-05)